Adding some test to evaluate the periodogram function

+++ test/utiltity.py
@@ -1,4 +1,5 @@
 import scipy.signal as sp
+import numpy as np
 from random import randint
 
 
@@ -11,7 +12,7 @@ def generate_inputs(number_inputs, minimum_size, maximum_size):
     data = []
     for i in range(0, number_inputs):
         size = randint(minimum_size, maximum_size)
-        data.append(signals[randint(0, len(signals) - 1)](size) + 0.1)
+        data.append(signals[randint(0, len(signals) - 1)](size).astype(np.float64) + 0.1)
     return data
 
 
@@ -21,6 +22,6 @@ def generate_pair_inputs(number_inputs, minimum_size, maximum_size):
     second = []
     for j in range(0, number_inputs):
         size = randint(minimum_size, maximum_size)
-        first.append(signals[randint(0, len(signals) - 1)](size) + 0.1)
-        second.append(signals[randint(0, len(signals) - 1)](size) + 0.1)
+        first.append(signals[randint(0, len(signals) - 1)](size).astype(np.float64) + 0.1)
+        second.append(signals[randint(0, len(signals) - 1)](size).astype(np.float64) + 0.1)
     return first, second
